+++ noodles-csi/src/lib.rs
@@ -18,7 +18,7 @@ pub use self::{
 };
 
 #[cfg(feature = "async")]
-pub use self::r#async::Reader as AsyncReader;
+pub use self::r#async::{Reader as AsyncReader, Writer as AsyncWriter};
 
 use std::{fs::File, io, path::Path};